// src/enums/ColumnsRequired.ts
export enum ColumnsRequired {
    TitleEn = 'Title_en',
    //FolderPath = 'FolderPath',
    FileDesc = 'FileDescription',
    DocumentTypeId = 'DocumentTypeId',
    DocumentType = 'DocumentType',
    PermittedFormats = 'PermittedFormats',
    UploadedFile = 'uploadedFile',
    Deadline = 'Deadline',
    Status = 'Status'
}