// src/enums/FileStatus.ts
export enum FileStatus {
    Processing = 'processing',
    Accepted = 'accepted',
    Declined = 'declined',
    Waiting = 'waiting',
    Requested = 'requested',
    Deleted = 'deleted',
    Published = 'published',
    AmlReturned = 'aml-returned',
    Provided = 'provided',
    PreparedForClient = 'prepared-for-client',
    Verification = 'progress-verification',
    Returned = 'returned',
    NotValid = 'not-valid',
    BoVerified = 'bo-verified',
    Prepared = 'prepared',
    BankRevision = 'bank-revision',
    ClientRevision = 'client-revision',
    FinalVersion = 'final-version',
    AcceptedConditional = 'accepted-conditional',
    NoType = 'no-type',
    Exported = 'exported',
    Signed = 'signed',
    ReturnedToSign = 'returned-to-sign',
    NotProvided = 'not-provided'
}