// src/enums/ColumnsDocTypes.ts
export enum ColumnsDocTypes {
    TitleEn = 'Title_en',
    UploadStatus = 'UploadStatus',
    Permissions = 'Permissions',
    FolderPath = 'FolderPath',
    FolderPathAfter = 'FolderPathAfter',
    ValidityStatuses = 'ValidityStatuses',
    FolderPathInternal = 'FolderPathInternal',
    FolderPathInternalAfter = 'FolderPathInternalAfter'
}